updating routes so after login, user will be redirected to accounts/ so :id in URL will not be displayed
