+++ views/account_new.ejs
@@ -16,7 +16,7 @@
 
     <main style="margin: 1em;">
       <h3>Create New Account</h3>
-      <form class="form-inline" action="/new" method="POST">
+      <form class="form-inline" action="/account/new" method="POST">
         <div class="form-group mb-2">
           <label for="url">URL:</label>
           <input class="form-control" type="text" name="url" placeholder="http://" style="width: 300px; margin: 1em">
